requirements.txt
streamlit
openai
python-dotenv
pydub
yt-dlp
pyTelegramBotAPI
